test-app-mobile bundling (#39)
Make metro bundling work using the test-app-mobile package. Add bundling to the CI loop to ensure that it continues to work.

I updated the metro build task so it always runs via the CLI (out-of-proc). There was an in-proc version of the task, but it was failing on a bad transformer, so I removed that for now. If we need in-proc bundling, we can bring the code back and get it working.

I added an option to override the 'assets' directory. The default behavior is to match the bundle directory. react-native-test-app uses this on android bundle runs, and set it to a subdir named 'res'.

I added package.json config for metro bundling, and updated the "scripts" commands for bundling to use rnx-kit-scripts. This ensures that our metro task runs and invokes the metro CLI correctly.

I updated the PR and Build CI loops to run bundling across the repo, ensuring that bundling errors break the build.