fix(dep-check): dev-only dependencies should always be added (#210)
The previous fix for dev-only dependencies (#205) had an unintentional
side-effect that adds `react-native-test-app` to `peerDependencies`.
This change makes it so that dev-only dependencies are always added.
This will allow test app only packages to still use
`react-native-test-app` even when `kitType: "app"`. It will also enable
type definition packages should we add them in the future.