fix: align on common console logger (#318)
Ideally, we should be using the reporter that is passed to Metro but we
don't have access to it from the `customSerializer` hook. The next best
thing would be to use
[`@react-native-community/cli-tools`](https://github.com/react-native-community/cli/blob/master/packages/tools/src/logger.ts),
but it includes a bunch of other tools and we cannot take a dependency
on any single version since we need to support multiple versions of
`react-native` (and hence multiple versions of
`@react-native-community/cli-tools`).

I've also looked briefly at other libraries, like
[`just-task-logger`](https://github.com/microsoft/just/tree/master/packages/just-task-logger),
but the purpose of this package is to match with the output of Metro.
`just-task-logger` in particular uses its own colour scheme and doesn't
allow configuring them to match Metro's colour scheme.