Adding esbuild plugin for mapping lib/ imports to src/ (#475)
This is similar to the babel-plugin-import-path remapper.

Having a monorepo of local packages that compiles from typescript source code in
`src/` to distributed javascript in `lib/`, this plugin rewrites all imports
from pointing to the generated artifacts to point directly to the source. This
makes it possible to consume local packages without first having to compile
them.