feat(cli): add `rnx-copy-assets` command
Adds a command to copy assets that are not referenced from JS. Usually,
Metro copies imported assets for you, but sometimes you need additional
files if they are only accessed from native modules.

The aim of this command is to generalize/standardize how to get native
assets into an app bundle. While this currently works for our internal
app, it is incomplete. As such, I'm avoiding advertising it in the
README for now.